ai_agent/agen_ai_multimodal/requirements.txt
phidata==2.7.2
streamlit==1.40.2
openai==0.28
transformers 
torch 
torchvision